Index classes are clickable! Posts does not work
posts are not clickable (or comments) -- there is something wrong because it is dynamic. Onclick does not work because it seems to be loops infinitely. PROBLEM: $(.subjectDiv).click(function () { ... } does not work yet (but should be used instead of list.setAttribute("onclick", ...).

Also deleted the html files we don't need.